// frontend/apps/web/src/domains/api/generated/model/questionCategory.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * PostPilot
 * OpenAPI spec version: 0.1.0
 */

export type QuestionCategory = typeof QuestionCategory[keyof typeof QuestionCategory];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const QuestionCategory = {
  technical: 'technical',
  behavioral: 'behavioral',
  system_design: 'system_design',
  role_specific: 'role_specific',
  culture_fit: 'culture_fit',
} as const;